Move the status property up to the root app shell view model
