+++ BlogMaui/AppShellViewModel.cs
@@ -1,6 +1,7 @@
 ﻿using BlogMaui.Authentication;
 using CommunityToolkit.Mvvm.ComponentModel;
 using CommunityToolkit.Mvvm.Input;
+using Jinaga;
 using Jinaga.Maui.Authentication;
 using MetroLog.Maui;
 using System.Windows.Input;
@@ -11,6 +12,7 @@ public partial class AppShellViewModel : ObservableObject
 {
     private readonly AuthenticationService authenticationService;
     private readonly UserProvider userProvider;
+    private readonly JinagaClient jinagaClient;
     private readonly LogController logController = new LogController();
 
     [ObservableProperty]
@@ -18,11 +20,14 @@ public partial class AppShellViewModel : ObservableObject
     [ObservableProperty]
     private string error = string.Empty;
 
+    [ObservableProperty]
+    private string status = string.Empty;
+
     public ICommand LogIn { get; }
     public ICommand LogOut { get; }
     public ICommand ViewLogs { get; }
 
-    public AppShellViewModel(AuthenticationService authenticationService, UserProvider userProvider)
+    public AppShellViewModel(AuthenticationService authenticationService, UserProvider userProvider, JinagaClient jinagaClient)
     {
         LogIn = new AsyncRelayCommand(HandleLogIn);
         LogOut = new AsyncRelayCommand(HandleLogOut);
@@ -30,8 +35,19 @@ public partial class AppShellViewModel : ObservableObject
 
         this.authenticationService = authenticationService;
         this.userProvider = userProvider;
+        this.jinagaClient = jinagaClient;
 
         logController.IsShakeEnabled = true;
+    }
+
+    public void Load()
+    {
+        jinagaClient.OnStatusChanged += JinagaClient_OnStatusChanged;
+    }
+
+    public void Unload()
+    {
+        jinagaClient.OnStatusChanged -= JinagaClient_OnStatusChanged;
     }
 
     private async Task HandleLogIn()
@@ -73,6 +89,36 @@ public partial class AppShellViewModel : ObservableObject
         catch (Exception ex)
         {
             Error = $"Error while logging out: {ex.GetMessage()}";
+        }
+    }
+
+    private void JinagaClient_OnStatusChanged(JinagaStatus status)
+    {
+        if ((!status.IsSaving || status.LastSaveError != null) && status.QueueLength > 0)
+        {
+            // There are facts in the queue, and
+            // the client is not saving, or has
+            // experienced an error.
+            Status = "Red";
+            if (status.LastSaveError != null)
+            {
+                Error = status.LastSaveError.GetMessage();
+            }
+        }
+        else if (status.IsSaving && status.QueueLength == 0)
+        {
+            // There was an error last time the client
+            // tried loading.
+            Status = "Yellow";
+            if (status.LastLoadError != null)
+            {
+                Error = status.LastLoadError.GetMessage();
+            }
+        }
+        else
+        {
+            Status = "Green";
+            Error = string.Empty;
         }
     }
 }
